python/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
motor==3.4.0
pymongo==4.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.27.0
pydantic-settings==2.2.1